Implementation of Baum-Welch Algorithm
The implementation of this algorithm that is displayed, is in python language. It is used to predict the stock movement of NSE (National Stock Exchange of India), and based on that movement, the user is suggested to buy or not to buy any stock.
The stock history is retrieved from the “nsepy” library. Here the class “HMM” defines the Hidden Markov Model and the function “baum_welch” implements the algorithm.